Organize code a bit
