+++ src/js/models/blocks/BlockState.js
