+++ src/js/views/NextPieceView.js
@@ -1,0 +1,16 @@
+import BlockRenderer from "./BlockRenderer";
+
+export default class NextPieceView extends BlockRenderer {
+    constructor(props) {
+        super(props);
+        this.state = {
+        }
+    }
+
+    renderBlocks() {
+    }
+
+    render() {
+    }
+  }
+
